adk_mh_mentor/requirements.txt
google-adk==0.5.0
google-cloud-aiplatform[adk, agent_engines]==1.88.0
uvicorn